actra-backend/requirements.txt
websockets>=12.0
redis[hiredis]>=5.0
google-genai>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
structlog>=24.0
httpx>=0.27.0
python-dotenv>=1.0
cartesia[websockets]>=1.0.0
pytest>=8.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
